pyproject.toml
[project]
name = "joshbduncan-blog"
version = "0.1.0"
requires-python = ">=3.13"
dependencies = [
    "beautifulsoup4>=4.13.3",
    "flask>=3.1.0",
    "flask-flatpages>=0.8.3",
    "markdown>=3.7",
    "pygments>=2.19.1",
    "python-dotenv>=1.0.1",
    "werkzeug>=3.1.3",
]

[dependency-groups]
dev = ["pyright>=1.1.396", "ruff>=0.9.9"]
